requirements.txt
matplotlib~=3.1.3
pyntcloud~=0.1.2
numpy~=1.18.1
pandas~=1.0.3
tqdm~=4.44.1
tensorflow~=1.15.0
tensorflow-compression==1.3
pyyaml~=5.1.2
pytest~=5.4.2
scipy~=1.4.1
numba~=0.49.0
open3d~=0.10.0
